fix: explicitly export interaction types
