+++ packages/http-framework/src/lib/interactions/index.ts
@@ -4,6 +4,7 @@ export * from './context-menu/decorators.js';
 export * from './context-menu/shared.js';
 export * from './resolvers/InteractionOptions.js';
 export * from './shared/decorators.js';
-export * from './structures/interactions/index.js';
 export * from './structures/Message.js';
-export { makeInteraction, type AbortError, type AddFiles, type DiscordError, type NonPingInteraction } from './utils/util.js';
+export * from './structures/interactions/index.js';
+export type * from './utils/util-types.js';
+export { makeInteraction } from './utils/util.js';
