+++ packages/http-framework/src/lib/interactions/structures/interactions/ModalSubmitInteraction.ts
@@ -1,8 +1,9 @@
 import { container } from '@sapphire/pieces';
 import { InteractionResponseType, Routes, type APIModalSubmitInteraction, type RESTPostAPIInteractionFollowupResult } from 'discord-api-types/v10';
-import { resultFromDiscord, type AsyncDiscordResult } from '../../utils/util.js';
-import { Data } from '../common/symbols.js';
+import type { AsyncDiscordResult } from '../../utils/util-types.js';
+import { resultFromDiscord } from '../../utils/util.js';
 import { Message, PartialMessage } from '../Message.js';
+import { Data } from '../common/symbols.js';
 import { BaseInteraction } from './base/BaseInteraction.js';
 import type {
 	DeferResponseData,
