+++ packages/http-framework/src/lib/interactions/structures/interactions/base/CommandInteraction.ts
@@ -7,7 +7,8 @@ import {
 	type APIUserApplicationCommandInteraction,
 	type RESTPostAPIInteractionFollowupResult
 } from 'discord-api-types/v10';
-import { resultFromDiscord, type AsyncDiscordResult } from '../../../utils/util.js';
+import type { AsyncDiscordResult } from '../../../utils/util-types.js';
+import { resultFromDiscord } from '../../../utils/util.js';
 import { Message, PartialMessage } from '../../Message.js';
 import { BaseInteraction } from './BaseInteraction.js';
 import type {
